+++ station/Route.cpp
@@ -1,11 +1,1 @@
 #include<Route.h>
-#include<iostream>
-#include <list>
-#include<functional>
-#include<StationIF.h>
-class Route
-{
-    public:
-        list<StationIF>stationif;
-
-};
