+++ station/RouteManager.cpp
@@ -1,16 +1,3 @@
 #include<RouteManager.h>
-#include<Route.h>
-#include<iostream>
-#include<functional>
-#include<list>
-using namespace std;
-class RouteManager
-{
-    public:
-        list<Route>routes; 
-        //void AddRoute();
-        void DeleteRoute();
-        void MakeFunctional();
 
-};
