+++ station/RouteManager.h
@@ -1,4 +1,14 @@
+#include<Route.h>
+#include<iostream>
+#include<functional>
+#include<list>
+using namespace std;
 class RouteManager
 {
+    public:
+        list<Route>routes; 
+        //void AddRoute();
+        void DeleteRoute();
+        void MakeFunctional();
 
 };
