+++ station/StationIF.cpp
@@ -1,8 +1,2 @@
-#include<iostream>
-#include<StationIF.h>
-using namespace std;
-class StationIF
-{
-    public:
 
-};
+#include<StationIF.h>
