+++ station/StationIF.h
@@ -1,3 +1,5 @@
+#include<iostream>
+using namespace std;
 class StationIF
 {
     public:
