+++ station/station.h
@@ -1,6 +1,12 @@
 #include<iostream>
 #include<StationIF.h>
-class Station : public StationIF
+#include<iostream>
+#include<string>
+using namespace std;
+class Station:public StationIF
 {
-   
+    public:
+    enum Station_Code{ Ahm ,Abu ,pali,kalol};
+    string stName;
+    bool IsFunctional;
 };
